+++ src/components/About/Divisions.tsx
